backend/pyproject.toml
[project]
name = "oxychat-backend"
version = "0.1.0"
description = "FastAPI backend service for OxyChat - AI workspace for meeting transcripts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.114.1,<0.116",
    "httpx>=0.28,<0.29",
    "uvicorn[standard]>=0.36,<0.37",
    "openai>=1.40",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "greenlet>=3.0.0",
    "chromadb>=0.4.0",
    "alembic>=1.17.2",
    "supabase>=2.27.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.4,<0.7",
    "mypy>=1.8,<2",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["chroma_data*", "*.egg-info*", ".venv*"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = ["I"]
